chore: move api types to /types
