+++ src/api.ts
@@ -1,7 +1,7 @@
 import createFetchClient from 'openapi-fetch';
 import createClient from 'openapi-react-query';
 import { QueryClient } from '@tanstack/react-query';
-import type { paths } from './api.d'; // generated by openapi-typescript
+import type { paths } from './types/api'; // generated by openapi-typescript (see README for details)
 import env from './env';
 
 const BACKEND_LOCATIONS_URL = env.VITE_API_URL === 'local' ? '/' : `${env.VITE_API_URL}`;
